requirements.txt
blinker==1.8.2
click==8.1.8
colorama==0.4.6
flask==3.0.3
Flask-Cors==5.0.0
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==2.1.5
pillow==10.4.0
werkzeug==3.0.6
zipp==3.20.2
discord.py==2.3.2
python-dotenv==1.0.0
